Generate a config.h file for versioning
Versioning was already handled in CMakeLists.txt, but it was handled
in a non-standard and unclear manner. We used
    set(VERSION
for starters. While this is valid CMake syntax, VERSION is also a
reserved word in cmake, depending on the context. We were
communicating the version information by a compile-time define, again,
not a recommended practice.

Change versioning to the more standard way of
    set(*VERSION_MAJOR
    set(*VERSION_MINOR
    set(*VERSION_MICRO
    set(*VERSION_STRING

Instead of defining the version at compile time, generate a config.h
file with the version information.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>