Unify the if(WIN32) defines
The if(WIN32)/else(WIN32)/endif(WIN32) is an archaic cmake syntax.
Remove all the if(WIN32) clauses, and use an option to specify whether
we want to link to the static versions of libsigrok*. That's basically
what the WIN32 clauses were doing anyway. This makes the "what is
going on" more readeable.

On WIN32, automatically set this option to link to the static libraries.