cmake: Add boost-thread as explicit dependency.
Apparently boost-thread requires boost-system even if not explicitly
used in the code.

See also: https://svn.boost.org/trac/boost/ticket/7241

On Debian with libboost-thread1.50.0 and libboost-system1.50.0 the
following error occurs during linking:

  /usr/bin/ld: CMakeFiles/pulseview.dir/main.cpp.o: undefined reference to symbol 'boost::system::system_category()'
  /usr/bin/ld: note: 'boost::system::system_category()' is defined in DSO /usr/lib/libboost_system.so.1.50.0 so try adding it to the linker command line
  /usr/lib/libboost_system.so.1.50.0: could not read symbols: Invalid operation
  collect2: ld returned 1 exit status

This commit fixes the issue.

It's unclear whether this is an issue that should be fixed in PulseView
in general, or whether it's a Boost issue.