buildsystem: Link against the platform's thread library
At least on Linux, boost-thread depends on pthread. The linking is not
automatic, and users have had issues with this.

We need to link against the platform's thread library, as it is needed for
boost-thread. CMake takes care of getting the exact settings to be passed
to the linker for us with ${CMAKE_THREAD_LIBS_INIT}.

This fixes bug #42.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>