Print --help/--version output to stdout (not stderr).
Thanks R. Diez <rdiezmail-comparevcd@yahoo.de> for the patch!

This is not a feasible practice for CLI tools where the output might
be piped into other tools (and you don't want to pipe help messages or
other non-data). However, for the PulseView GUI this is acceptable since
it's not meant to be used that way.