decode: Clear error message upon every new session run.
When a decoder throws an exception, i.e. srd_session_send() returns an error,
PulseView shows a red "Failed to initialise decoder" banner in the
decode trace at the moment.

This can happen when e.g. invalid option values are set in the PD popup,
or invalid probe combinations, and so on. However, after the user resolved
the error condition (e.g. selected a proper set of probes / options) the
red error banner would still remain visible and no re-decoding would happen.

This patch fixes that.