Fix bug #235. Add shortcut keys for View/Zoom In (Ctrl++) and View/Zoom Out (Ctrl+-)
These are the shortcut keys for zoom in and out as suggested by
QT. Works with the normal keyboard and numpad keys.

Seems that the preferred simple + and - keys have an issue.
While the - key is working and displayed in the menu as expected,
the + key is zooming but not displayed at View/Zoom In
in the menu. I've tried several attempts.