CMakeLists.txt: Pass -DENABLE_SIGNALS to the compiler.
Until now only the cmake variable ENABLE_SIGNALS was used to add
signalhandler.{cpp,h} to the list of files to build.

However, main.cpp also uses '#ifdef ENABLE_SIGNALS' which didn't work
so far, since -DENABLE_SIGNALS was not passed to the compiler.

(one effect being that PulseView cannot be killed via CTRL-C from
an xterm or the like)

This fixes bug #368.