Use QApplication metadata for persistent storage
Since the application metadata is set in application.c,
we should use it instead of providing it manually every
time we create a QSettings object.

As I understand it, subclassing QApplication in application.c
is only a workaround for bad Qt behavior. In the event
that it is removed, the application metadata should be moved
so that persistent storage still works.