connect: Fix filtering of supported devices.
Now that driver options are split from device options, it is not
possible to check the SAMPLERATE option on a driver as it is a device
only option. The SAMPLERATE test still works on some drivers that
didn't have their options split yet, but this won't hold true for long.
The proper way is to check the driver class, which is the only option
you can expect at the driver level for now.