MainBar: Prevent context menu from appearing
A user can remove the main toolbar by using the right-click context
menu and clicking on the single entry in it, effectively hiding the
toolbar this way. There is no way to restore it except for closing
PV and deleting the stored session file in ~/.config/sigrok.

This patch prevents the context menu from appearing, thereby
removing the ability to hide the main tool bar.