Bring back sigrok-logo-notext.png (needed on Android).
On Android sigrok-logo-notext.png is used as app icon, and the build
thus fails if it is not available.

Changing android/custom_rules.xml to use the SVG version instead doesn't
seem to work. It appears Android doesn't really support SVGs as app
icons, or at least not without further changes. Continue using the PNG
for now to keep the Android build in a working state.