Untangle the time formatting functions a bit
The formatting depending on the distance between two timestamps is
reintroduced with the responding function residing in the 'Ruler' class
(the prime user of that function).

Fixes a rounding bug for the least significant digit in the
'format_time_minutes()' function.

The 'Cursor' and 'CursorPair' classes now use the same precision when
formatting timestamps as the rest of the program.