DecodeTrace: Improve annotation block drawing
Instead of just drawing blocks using the color of the first
annotation in the block, draw a "special" annotation
with a dotted pattern.
The color of this annotation depends on the annotations
in that range: a color is used if they're all of the same
annotation class. If there's different kinds, gray is used.