DecodeTrace: Prevent trace height from jumping
In 506317984393a41000650ccca868f3dd66872a4f, the decode
trace height was made dependent on the number of currently
visible annotation rows.
This however has issues: when zooming in on an area where
there are no annotations, the height of the decode trace
collapses. With this patch, this can't happen anymore as
a new variable keeps track of the maximum number of
annotation rows ever encountered, keeping the decoder
trace height stable.