DecoderStack: Force a viewport update more frequently
Updating the viewport every 65kib of input data causes
the decode trace to make really big leaps, leaving the
user with the feeling of a not-so-snappy user interface.
We can easily change this by updating the trace every 1kib
since the decoder and the UI thread are separate and the
decode trace drawing is much faster since when the limit
was originally set.