cmake, INSTALL: Bump the Boost requirement to >= 1.55.
Support for serialization of multiprecision data (which is used for view
timestamps) was introduced in boost version 1.55. Since commit 3a21afa6
builds with versions up to and including 1.54 are reported to fail.

Bump the required version from 1.53 to 1.55. Rephrase the CMake logic
which searches for the Boost dependency. Create a list of libraries from
fixed and optional parts, which as a byproduct avoids long lines. Pass
that list of required libraries to one common find_package() call.

Signed-off-by: Gerhard Sittig <gerhard.sittig@gmx.net>