DecodeSignal: Change srd session handling
Before, we would hold onto the global srd mutex until all data
was processed and then destroy the srd session, making it
necessary to re-create the srd session (and thus run all data
through the PD again) when new data comes in.

This patch changes this so that we keep the srd session around
when we wait for input data, allowing us to re-use the existing
session when new data comes in. Also, this enables us to
release the global srd mutex when waiting, which wasn't
possible before.