DecodeSignal: Adjust the DecodeChunkLength
The initial 4096 was too low, though it did make the UI
quite responsive. In 9cc264 it was increased to 10*1024*1024
to speed up decoding but that makes the UI stall very badly
under certain circumstances (e.g. when changing the baud rate
for UART by entering digits).
The reason is that every time a decoder property is changed,
the decode is restarted. Restarting the decode means that
the currently running processes need to shut down, so
they're told to do so and the GUI thread waits for them
to finish.
However, the srd_session_send() call is blocking, meaning
that libsrd will happily block the PV GUI thread until it's
done processing the current chunk of data.

For this reason, DecodeChunkLength can't be too high and
256*1024 looks like a reasonable compromise for now.