View: Improve 2a9fcd621 by using settings_restored_
The timer is no longer needed as the introduction of
settings_restored_ allows us to determine whether we
should resize the trace label header to its proper
width when the show event is received.
This is because the show event is received only after
all widget resizing took place. This means that the
header pane sizes will be reliable at this point,
making this solution much cleaner than relying on
a timer repeatedly calling expand_header_to_fit().