DecodeSignal: Don't assume channels always have logic data
It could be that a signal was assigned to a decoder when it
had a conversion enabled, then the conversion was removed.
This means the signal may no longer provide logic data but
as it's still assigned to the decoder, it will try to access
the logic data and... boom.