DecodeSignal: Don't set name when creating the signal
The thought behind this was to have no unnamed signals because
this may make debugging harder. However, giving the signal a
long name and shortening it when decoders are added isn't working
well. The reason is that the width of the header area can be
expanded programmatically but we don't shrink it because we
don't want to hide things the user may want to see.
This behavior in turn leads to the header width expanding when
a signal labeled "Empty decoder signal" is added, only to see
this very signal be renamed to e.g. "SPI" immediately afterwards.
As we don't shrink the header width, the header is now too wide.