Disable unit tests by default
The rationale for this change is that except for developers, pretty
much everyone who builds from source does so because they want to
use the latest changes. However, only developers care about the results
of the unit tests, so this doubles the build time for users without
any real benefit.

Even worse, the unit tests are known to fail on several hosts for
reasons that are within the realm of the unit tests themselves, not
because of broken code. This confuses users and even prevents them
from installing PV for no good reason.

Since the unit tests do still run on Jenkins, there's little to no
coverage gap in my opinion.